moved documentSymbols into AnalysisBase
