+++ src/powerquery-language-services/analysis/analysis.ts
@@ -2,7 +2,14 @@
 // Licensed under the MIT license.
 
 import { CommonError, Result } from "@microsoft/powerquery-parser";
-import type { FoldingRange, Hover, Location, Position, SignatureHelp } from "vscode-languageserver-types";
+import type {
+    DocumentSymbol,
+    FoldingRange,
+    Hover,
+    Location,
+    Position,
+    SignatureHelp,
+} from "vscode-languageserver-types";
 
 import { AutocompleteItem } from "../inspection";
 import { IDisposable } from "../commonTypes";
@@ -12,6 +19,7 @@ import { TextEdit } from "vscode-languageserver-textdocument";
 export interface Analysis extends IDisposable {
     getAutocompleteItems(position: Position): Promise<Result<AutocompleteItem[] | undefined, CommonError.CommonError>>;
     getDefinition(position: Position): Promise<Result<Location[] | undefined, CommonError.CommonError>>;
+    getDocumentSymbols(): Promise<Result<DocumentSymbol[] | undefined, CommonError.CommonError>>;
     getFoldingRanges(): Promise<Result<FoldingRange[] | undefined, CommonError.CommonError>>;
     getHover(position: Position): Promise<Result<Hover | undefined, CommonError.CommonError>>;
     getPartialSemanticTokens(): Promise<Result<PartialSemanticToken[] | undefined, CommonError.CommonError>>;
