+++ src/powerquery-language-services/analysis/analysisBase.ts
@@ -7,8 +7,24 @@ import {
     AstNodeById,
     ChildIdsById,
 } from "@microsoft/powerquery-parser/lib/powerquery-parser/parser/nodeIdMap/nodeIdMap";
-import type { FoldingRange, Hover, Location, Position, SignatureHelp } from "vscode-languageserver-types";
-import { NodeIdMap, ParseError, ParseState, TXorNode } from "@microsoft/powerquery-parser/lib/powerquery-parser/parser";
+import {
+    DocumentSymbol,
+    FoldingRange,
+    Hover,
+    Location,
+    Position,
+    SignatureHelp,
+    SymbolKind,
+} from "vscode-languageserver-types";
+import {
+    NodeIdMap,
+    NodeIdMapUtils,
+    ParseError,
+    ParseState,
+    TXorNode,
+    XorNode,
+    XorNodeUtils,
+} from "@microsoft/powerquery-parser/lib/powerquery-parser/parser";
 import type { TextDocument, TextEdit } from "vscode-languageserver-textdocument";
 import { Trace, TraceConstant } from "@microsoft/powerquery-parser/lib/powerquery-parser/common/trace";
 import { Ast } from "@microsoft/powerquery-parser/lib/powerquery-parser/language";
@@ -24,7 +40,7 @@ import type {
     PartialSemanticToken,
     SignatureProviderContext,
 } from "../providers/commonTypes";
-import { CommonTypesUtils, Inspection, InspectionSettings, PositionUtils } from "..";
+import { CommonTypesUtils, Inspection, InspectionSettings, InspectionUtils, PositionUtils } from "..";
 import {
     findDirectUpperScopeExpression,
     findScopeItemByLiteral,
@@ -169,6 +185,25 @@ export abstract class AnalysisBase implements Analysis {
             }
 
             return result.value;
+        }, this.analysisSettings.inspectionSettings.locale);
+    }
+
+    public getDocumentSymbols(): Promise<PQP.Result<DocumentSymbol[] | undefined, CommonError.CommonError>> {
+        return ResultUtils.ensureResultAsync(async () => {
+            const maybeParseState: ParseState | undefined = await this.getParseState();
+
+            if (maybeParseState === undefined) {
+                return undefined;
+            }
+
+            const nodeIdMapCollection: NodeIdMap.Collection = maybeParseState.contextState.nodeIdMapCollection;
+            const currentSymbols: DocumentSymbol[] = [];
+            const parentSymbolById: Map<number, DocumentSymbol> = new Map();
+
+            addIdentifierPairedExpressionSymbols(nodeIdMapCollection, currentSymbols, parentSymbolById);
+            addRecordSymbols(nodeIdMapCollection, currentSymbols, parentSymbolById);
+
+            return currentSymbols;
         }, this.analysisSettings.inspectionSettings.locale);
     }
 
@@ -998,4 +1033,119 @@ export abstract class AnalysisBase implements Analysis {
 
         return this.parseError;
     }
+}
+
+function addIdentifierPairedExpressionSymbols(
+    nodeIdMapCollection: NodeIdMap.Collection,
+    currentSymbols: DocumentSymbol[],
+    parentSymbolById: Map<number, DocumentSymbol>,
+): void {
+    const identifierPairedExpressionIds: Set<number> =
+        nodeIdMapCollection.idsByNodeKind.get(Ast.NodeKind.IdentifierPairedExpression) ?? new Set();
+
+    for (const nodeId of identifierPairedExpressionIds) {
+        const xorNode: XorNode<Ast.IdentifierPairedExpression> =
+            NodeIdMapUtils.assertGetXorChecked<Ast.IdentifierPairedExpression>(
+                nodeIdMapCollection,
+                nodeId,
+                Ast.NodeKind.IdentifierPairedExpression,
+            );
+
+        if (!XorNodeUtils.isAstXor(xorNode)) {
+            continue;
+        }
+
+        const astNode: Ast.IdentifierPairedExpression = xorNode.node;
+        const documentSymbol: DocumentSymbol = InspectionUtils.getSymbolForIdentifierPairedExpression(astNode);
+        addDocumentSymbols(nodeIdMapCollection, parentSymbolById, nodeId, currentSymbols, documentSymbol);
+        parentSymbolById.set(nodeId, documentSymbol);
+    }
+}
+
+function addRecordSymbols(
+    nodeIdMapCollection: NodeIdMap.Collection,
+    currentSymbols: DocumentSymbol[],
+    parentSymbolById: Map<number, DocumentSymbol>,
+): void {
+    const recordIdCollections: ReadonlyArray<Set<number>> = [
+        nodeIdMapCollection.idsByNodeKind.get(Ast.NodeKind.RecordExpression) ?? new Set(),
+        nodeIdMapCollection.idsByNodeKind.get(Ast.NodeKind.RecordLiteral) ?? new Set(),
+    ];
+
+    for (const collection of recordIdCollections) {
+        for (const nodeId of collection) {
+            const xorNode: XorNode<Ast.RecordExpression | Ast.RecordLiteral> = NodeIdMapUtils.assertGetXorChecked<
+                Ast.RecordExpression | Ast.RecordLiteral
+            >(nodeIdMapCollection, nodeId, [Ast.NodeKind.RecordExpression, Ast.NodeKind.RecordLiteral]);
+
+            if (!XorNodeUtils.isAstXor(xorNode)) {
+                continue;
+            }
+
+            const astNode: Ast.RecordExpression | Ast.RecordLiteral = xorNode.node;
+
+            // Process the record if the immediate parent is a Struct
+            const parentId: number | undefined = nodeIdMapCollection.parentIdById.get(astNode.id);
+            const parentSymbol: DocumentSymbol | undefined = parentId ? parentSymbolById.get(parentId) : undefined;
+
+            if (parentSymbol && parentSymbol.kind === SymbolKind.Struct) {
+                const fieldSymbols: ReadonlyArray<DocumentSymbol> = InspectionUtils.getSymbolsForRecord(astNode);
+
+                if (fieldSymbols.length > 0) {
+                    addDocumentSymbols(
+                        nodeIdMapCollection,
+                        parentSymbolById,
+                        astNode.id,
+                        currentSymbols,
+                        ...fieldSymbols,
+                    );
+                }
+            }
+        }
+    }
+}
+
+function addDocumentSymbols(
+    nodeIdMapCollection: NodeIdMap.Collection,
+    parentSymbolById: Map<number, DocumentSymbol>,
+    nodeId: number,
+    currentSymbols: DocumentSymbol[],
+    ...newSymbols: DocumentSymbol[]
+): void {
+    const parentSymbol: DocumentSymbol | undefined = findParentSymbol(nodeIdMapCollection, parentSymbolById, nodeId);
+
+    if (parentSymbol) {
+        if (!parentSymbol.children) {
+            parentSymbol.children = [];
+        }
+
+        parentSymbol.children.push(...newSymbols);
+
+        return;
+    }
+
+    // Add to the top level
+    currentSymbols.push(...newSymbols);
+}
+
+function findParentSymbol(
+    nodeIdMapCollection: NodeIdMap.Collection,
+    parentSymbolById: Map<number, DocumentSymbol>,
+    nodeId: number,
+): DocumentSymbol | undefined {
+    // Get parent for current node
+    const parentNodeId: number | undefined = nodeIdMapCollection.parentIdById.get(nodeId);
+
+    if (!parentNodeId) {
+        // No more parents to check
+        return undefined;
+    }
+
+    let parentSymbol: DocumentSymbol | undefined = parentSymbolById.get(parentNodeId);
+
+    if (!parentSymbol) {
+        parentSymbol = findParentSymbol(nodeIdMapCollection, parentSymbolById, parentNodeId);
+    }
+
+    return parentSymbol;
 }
