+++ src/powerquery-language-services/analysis/analysisUtils.ts
@@ -12,7 +12,7 @@ const analysisByUri: Map<string, [Analysis, number]> = new Map();
 export function createAnalysis(document: TextDocument, analysisSettings: AnalysisSettings): Analysis {
     const cacheKey: string = document.uri.toString();
 
-    if (analysisByUri.has(cacheKey)) {
+    if (analysisSettings.isWorkspaceCacheAllowed && analysisByUri.has(cacheKey)) {
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         const [analysis, version]: [Analysis, number] = analysisByUri.get(cacheKey)!;
 
@@ -24,7 +24,10 @@ export function createAnalysis(document: TextDocument, analysisSettings: Analysi
     }
 
     const analysis: Analysis = new DocumentAnalysis(document, analysisSettings);
-    analysisByUri.set(cacheKey, [analysis, document.version]);
+
+    if (analysisSettings.isWorkspaceCacheAllowed) {
+        analysisByUri.set(cacheKey, [analysis, document.version]);
+    }
 
     return analysis;
 }
