+++ src/test/documentSymbols.ts
@@ -3,9 +3,9 @@
 
 import "mocha";
 import { assert, expect } from "chai";
+import { Assert, CommonError, Result } from "@microsoft/powerquery-parser";
 
-import * as DocumentSymbols from "../powerquery-language-services/documentSymbols";
-import { SymbolKind, TextDocument } from "../powerquery-language-services";
+import { Analysis, AnalysisUtils, DocumentSymbol, SymbolKind, TextDocument } from "../powerquery-language-services";
 import { TestConstants, TestUtils } from ".";
 import { AbridgedDocumentSymbol } from "./testUtils";
 import { MockDocument } from "./mockDocument";
@@ -15,9 +15,13 @@ async function expectSymbolsForDocument(
     document: TextDocument,
     expectedSymbols: ReadonlyArray<AbridgedDocumentSymbol>,
 ): Promise<void> {
-    const actualSymbols: ReadonlyArray<AbridgedDocumentSymbol> = TestUtils.createAbridgedDocumentSymbols(
-        await DocumentSymbols.getDocumentSymbols(document, TestConstants.SimpleInspectionSettings, false),
-    );
+    const analysis: Analysis = AnalysisUtils.createAnalysis(document, TestConstants.SimpleLibraryAnalysisSettings);
+
+    const actualSymbols: Result<DocumentSymbol[] | undefined, CommonError.CommonError> =
+        await analysis.getDocumentSymbols();
+
+    Assert.isOk(actualSymbols);
+    Assert.isDefined(actualSymbols.value);
 
     assert.isDefined(actualSymbols);
